Create app module to remove the event listening logic to a separate module
